Implemented GNSS warmup mode
The GNSS can now be „warmed up“, meaning it is powered without the
tracker running (and transmitting packets!). This should help in cases
where the GNSS module is still in an unrealiable state and every
transmission causes it to loose the position again.

The new mode can also be used in combination with the receiver: In case
you want to continously track other stations, switch on the GPS warmup
and distance and heading will always be accurate for received packets.

Unfortunately, it was not possible to manage the GNSS via the periph_pwr
module, so a new global flag was introduced. If the tracker or the
warmup mode is active, the GPS is powered, otherwise it is off.