bme280: power management integration
Unfortunately, the BME280 is also connected to the PWR switch which
means that the entire system has to be switched on for the readout :/

Additionally, it means that we have to introduce a startup delay which
is 2 ms starting at VDD ramp-up, according to the datasheet. We wait 5
ms after enabling the peripheral power to be on the safe side.