main: prevent crash during settings initialization
The crash can happen if the FDS subsystem cannot be initialized because
too few empty flash pages are available. In that case, the firmware
tries to erase the flash pages intended for settings storage and calls
nrfx_nvmc_page_erase() to do so.

The problem is that this function is called after the SoftDevice (BLE
stack) has been initialized, which causes a hard fault (the SoftDevice
reports an invalid memory access).

To solve this issue, the settings module is now initialized before the
BLE stack.