DAO: use human-readable version
This version of DAO (identified by 'W') just adds an additional digit to the
already human-readable location format. Thus, the whole packet is now
human-readable.

Note that precision does not suffer by this change because the coordinates are
stored as 32-bit float internally which limits the resolution to 40000 km /
2^24 = 2.38m. W-DAO gives a resolution of 40000 km / (360 (deg) * 60 (min) *
1000 (fract. min)) = 1.85m.

This change was suggested and supplied by Thomas DL9SAU.