aprs: use "destination call digipeating"
This change saves airtime by replacing the "WIDEn-n" digipeating in the path
with so-called "destination call digipeating". This technique puts the n from
WIDEn-n into the APRS destination call as SSID. This makes the addressing quite
compact.

Additionally, digipeating is completely disabled for weather reports.

This change was proposed by Thomas DL9SAU.