Try to properly shut down LoRa module
The SX1262 always needs to go to STBD_RC before it can go to SLEEP mode
(which is the proper power saving mode). This is ensured now, even if
the receiver is active.

This is/was a problem if the module is wired such that it is always
powered. If the receiver was active, the firmware tried to put it
directly into SLEEP mode. Accordings to the datasheet, this does not
work, so the module stayed in RX mode while the firmware assumed it was
in SLEEP mode. Therefore, the module never actually went to sleep and
wasted energy.